Added new Fonts component and support for he new Fonts Library
