+++ inc/Fonts/Component.php
@@ -1,0 +1,205 @@
+<?php
+/**
+ * WP_Rig\WP_Rig\Base_Support\Component class
+ *
+ * @package wp_rig
+ */
+
+namespace WP_Rig\WP_Rig\Fonts;
+
+use WP_Rig\WP_Rig\Component_Interface;
+use WP_Rig\WP_Rig\Templating_Component_Interface;
+
+/**
+ * Class for adding basic theme support, most of which is mandatory to be implemented by all themes.
+ *
+ * Exposes template tags:
+ * * `wp_rig()->get_version()`
+ * * `wp_rig()->get_asset_version( string $filepath )`
+ */
+class Component implements Component_Interface, Templating_Component_Interface {
+
+	/**
+	 * Associative array of Google Fonts to load, as $font_name => $font_variants pairs.
+	 *
+	 * Do not access this property directly, instead use the `get_google_fonts()` method.
+	 *
+	 * @var array
+	 */
+	protected $google_fonts;
+
+	/**
+	 * Gets the unique identifier for the theme component.
+	 *
+	 * @return string Component slug.
+	 */
+	public function get_slug(): string {
+		return 'base_support';
+	}
+
+	/**
+	 * Adds the action and filter hooks to integrate with WordPress.
+	 */
+	public function initialize(): void {
+		add_action( 'wp_enqueue_scripts', array( $this, 'action_enqueue_fonts' ) );
+		add_action( 'after_setup_theme', array( $this, 'action_add_editor_fonts' ) );
+		add_action( 'init', array( $this, 'wprig_register_fonts') );
+		add_filter( 'wp_resource_hints', array( $this, 'filter_resource_hints' ), 10, 2 );
+	}
+
+	/**
+	 * Gets template tags to expose as methods on the Template_Tags class instance, accessible through `wp_rig()`.
+	 *
+	 * @return array Associative array of $method_name => $callback_info pairs. Each $callback_info must either be
+	 *               a callable or an array with key 'callable'. This approach is used to reserve the possibility of
+	 *               adding support for further arguments in the future.
+	 */
+	public function template_tags(): array {
+		return array(
+			'get_version'       => array( $this, 'get_version' ),
+			'get_asset_version' => array( $this, 'get_asset_version' ),
+		);
+	}
+
+	/**
+	 * Registers font collections with WordPress if the wp_register_font_collection function exists.
+	 *
+	 * @return void
+	 */
+	function wprig_register_fonts(): void {
+		if ( function_exists( 'wp_register_font_collection' ) ) {
+			wp_register_font_collection( 'google-fonts', [
+				[
+					'family' => 'Roboto Condenses',
+					'variants' => [ '300', '400', '500', '700' ],
+					'subsets' => [ 'latin', 'latin-ext' ],
+				],
+				[
+					'family' => 'Crimson Text',
+					'variants' => [ '300', '400', '600', '700' ],
+					'subsets' => [ 'latin', 'latin-ext' ],
+				],
+			] );
+
+			wp_register_font_collection( 'local-fonts', [
+				[
+					'family' => 'My Local Font',
+					'file' => get_template_directory_uri() . '/assets/fonts/my-local-font/my-local-font.woff2',
+					'weight' => '400',
+					'style' => 'normal',
+				],
+			] );
+		}
+	}
+
+	/**
+	 * Enqueues Google Fonts for the theme.
+	 *
+	 * @return void
+	 */
+	public function action_enqueue_fonts(): void {
+		// Enqueue Google Fonts.
+		$google_fonts_url = $this->get_google_fonts_url();
+		if ( ! empty( $google_fonts_url ) ) {
+			wp_enqueue_style( 'wp-rig-fonts', $google_fonts_url, array(), null ); // phpcs:ignore WordPress.WP.EnqueuedResourceParameters.MissingVersion
+		}
+
+	}
+
+	/**
+	 * Enqueues the Google Fonts for the WordPress editor.
+	 *
+	 * Retrieves the URL for the Google Fonts and, if it is not empty,
+	 * adds the editor styles to the WordPress editor.
+	 *
+	 * @return void
+	 */
+	public function action_add_editor_fonts(): void {
+		// Enqueue Google Fonts.
+		$google_fonts_url = $this->get_google_fonts_url();
+		if ( ! empty( $google_fonts_url ) ) {
+			add_editor_style( $this->get_google_fonts_url() );
+		}
+	}
+
+	/**
+	 * Adds preconnect resource hint for Google Fonts.
+	 *
+	 * @param array  $urls          URLs to print for resource hints.
+	 * @param string $relation_type The relation type the URLs are printed.
+	 * @return array URLs to print for resource hints.
+	 */
+	public function filter_resource_hints( array $urls, string $relation_type ): array {
+		if ( 'preconnect' === $relation_type && wp_style_is( 'wp-rig-fonts', 'queue' ) ) {
+			$urls[] = array(
+				'href' => 'https://fonts.gstatic.com',
+				'crossorigin',
+			);
+		}
+
+		return $urls;
+	}
+
+	/**
+	 * Returns Google Fonts used in theme.
+	 *
+	 * @return array Associative array of $font_name => $font_variants pairs.
+	 */
+	protected function get_google_fonts(): array {
+		if ( is_array( $this->google_fonts ) ) {
+			return $this->google_fonts;
+		}
+
+		$google_fonts = array(
+			'Roboto Condensed' => array( '400', '400i', '700', '700i' ),
+			'Crimson Text'     => array( '400', '400i', '600', '600i' ),
+		);
+
+		/**
+		 * Filters default Google Fonts.
+		 *
+		 * @param array $google_fonts Associative array of $font_name => $font_variants pairs.
+		 */
+		$this->google_fonts = (array) apply_filters( 'wp_rig_google_fonts', $google_fonts );
+
+		return $this->google_fonts;
+	}
+
+	/**
+	 * Returns the Google Fonts URL to use for enqueuing Google Fonts CSS.
+	 *
+	 * Uses `latin` subset by default. To use other subsets, add a `subset` key to $query_args and the desired value.
+	 *
+	 * @return string Google Fonts URL, or empty string if no Google Fonts should be used.
+	 */
+	protected function get_google_fonts_url(): string {
+		$google_fonts = $this->get_google_fonts();
+
+		if ( empty( $google_fonts ) ) {
+			return '';
+		}
+
+		$font_families = array();
+
+		foreach ( $google_fonts as $font_name => $font_variants ) {
+			if ( ! empty( $font_variants ) ) {
+				if ( ! is_array( $font_variants ) ) {
+					$font_variants = explode( ',', str_replace( ' ', '', $font_variants ) );
+				}
+
+				$font_families[] = $font_name . ':' . implode( ',', $font_variants );
+				continue;
+			}
+
+			$font_families[] = $font_name;
+		}
+
+		$query_args = array(
+			'family'  => implode( '|', $font_families ),
+			'display' => 'swap',
+		);
+
+		return add_query_arg( $query_args, 'https://fonts.googleapis.com/css' );
+	}
+
+}
